Clean up and minor rearrangement of the files.  Changes to be committed: upload_router_huggingface.py moved inside uploader_routers folder, and needed changes on the __init__ files and main.py were made. 	modified:   src/main.py 	modified:   src/routers/__init__.py 	new file:   src/routers/uploader_router.py 	modified:   src/routers/uploader_routers/__init__.py 	renamed:    src/routers/upload_router_huggingface.py -> src/routers/uploader_routers/upload_router_huggingface.py 	modified:   src/routers/uploader_routers/upload_router_zenodo.py 	modified:   src/tests/uploader/zenodo/test_dataset_zenodo_uploader.py uploader folder renamed to uploaders after having zenodo_uploader file created 	renamed:    src/uploader/__init__.py -> src/uploaders/__init__.py 	renamed:    src/uploader/hugging_face_uploader.py -> src/uploaders/hugging_face_uploader.py 	renamed:    src/uploader/utils.py -> src/uploaders/utils.py 	renamed:    src/uploader/zenodo_uploader.py -> src/uploaders/zenodo_uploader.py
