+++ src/routers/__init__.py
@@ -1,5 +1,1 @@
 from .resource_router import ResourceRouter  # noqa:F401
-from .upload_router_huggingface import UploadRouterHuggingface
-
-
-other_routers = [UploadRouterHuggingface()]
