+++ src/routers/uploader_routers/__init__.py
@@ -1,4 +1,5 @@
-# from .upload_router__huggingface import UploadRouterHuggingface
+from routers.uploader_router import UploaderRouter
 from .upload_router_zenodo import UploadRouterZenodo
+from .upload_router_huggingface import UploadRouterHuggingface
 
-router_list = [UploadRouterZenodo()]
+router_list: list[UploaderRouter] = [UploadRouterZenodo(), UploadRouterHuggingface()]
